09-Event-Grid/Python/demo-Function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core
azure-common
azure-functions
azure-data-tables
dotenv